+++ Tests/ReducerTests.swift
@@ -15,18 +15,18 @@ class MockReducerContainer {
     var reducer: Reducer<CounterState>!
 
     init() {
-        reducer = Reducer { action, state in
+        reducer = { action, state in
             self.calledWithAction.append(action)
             return state
         }
     }
 }
 
-let increaseByOneReducer = Reducer { action, state in
+let increaseByOneReducer: Reducer<CounterState> = { action, state in
     CounterState(count: state.count + 1)
 }
 
-let increaseByTwoReducer = Reducer { action, state in
+let increaseByTwoReducer: Reducer<CounterState> = { action, state in
     CounterState(count: state.count + 2)
 }
 
@@ -39,9 +39,11 @@ class ReducerTests: XCTestCase {
         let mockReducer1 = MockReducerContainer()
         let mockReducer2 = MockReducerContainer()
 
-        let combinedReducer = Reducer(mockReducer1.reducer, mockReducer2.reducer)
+        let combinedReducer: Reducer<CounterState> = { action, state in
+            mockReducer2.reducer(action, mockReducer1.reducer(action, state))
+        }
 
-        _ = combinedReducer.transform(NoOpAction(), CounterState())
+        _ = combinedReducer(NoOpAction(), CounterState())
 
         XCTAssertEqual(mockReducer1.calledWithAction.count, 1)
         XCTAssertEqual(mockReducer2.calledWithAction.count, 1)
@@ -53,9 +55,11 @@ class ReducerTests: XCTestCase {
      it combines the results from each individual reducer correctly
      */
     func testCombinesReducerResults() {
-        let combinedReducer = Reducer(increaseByOneReducer, increaseByTwoReducer)
+        let combinedReducer: Reducer<CounterState> = { action, state in
+            increaseByTwoReducer(action, increaseByOneReducer(action, state))
+        }
 
-        let newState = combinedReducer.transform(NoOpAction(), CounterState())
+        let newState = combinedReducer(NoOpAction(), CounterState())
 
         XCTAssertEqual(newState.count, 3)
     }